Updated ui positioning code and fixed damping option bug
