+++ src/GravitySimulation/GravitySimulation/GravitySimulation.vcxproj.filters
@@ -54,6 +54,9 @@
     <ClCompile Include="quadtree\node.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="ui\menu_utils.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resource.h">
